Add new deprecated useFetchState
