+++ frontend/src/pages/modelRegistry/screens/ModelVersionDetails/ModelVersionDetailsTabs.tsx
@@ -4,7 +4,8 @@ import { PageSection, Tab, Tabs, TabTitleText } from '@patternfly/react-core';
 import '~/pages/pipelines/global/runs/GlobalPipelineRunsTabs.scss';
 import { InferenceServiceKind, ServingRuntimeKind } from '~/k8sTypes';
 import { ModelVersion } from '~/concepts/modelRegistry/types';
-import { FetchStateObject, ListWithNonDashboardPresence } from '~/types';
+import { ListWithNonDashboardPresence } from '~/types';
+import { FetchStateObject } from '~/utilities/useFetch';
 import { SupportedArea, useIsAreaAvailable } from '~/concepts/areas';
 import { ModelVersionDetailsTabTitle, ModelVersionDetailsTab } from './const';
 import ModelVersionDetailsView from './ModelVersionDetailsView';
